Move overload of literal_unroll to avoid circular dependency that breaks Python 2.7
