+++ numba/targets/base.py
@@ -262,7 +262,7 @@ class BaseContext(object):
         # Populate built-in registry
         from . import (arraymath, enumimpl, iterators, linalg, numbers,
                        optional, polynomial, rangeobj, slicing, tupleobj,
-                       gdb_hook, hashing, heapq, literal_dispatch)
+                       gdb_hook, hashing, heapq, literal)
         try:
             from . import npdatetime
         except NotImplementedError:
